Streaks in the focal plane bug fix
The dm_map created by prop_dm is oversized compared to the input beam when using the actuator spacing arguments in the adaptive optics example in the guide. When this map is applied to the beam it leaves a discontuintiy on one edge if there's a slight misalaignment, or a thin ring if the alignment is perfect. To counter this a smooth gradial transition circular mask is applied to the output of the AO (apodising the pupil plane)

The same WFS map measured at the shortest wavelength is applied (after spatial scaling to beam width) to all wavelengths (second dimension os CPA_maps is now redudant)